# Use parallel gzip