… vendor/src/github.com/klauspost/compress/LICENSE
… vendor/src/github.com/klauspost/compress/flate/copy.go
… vendor/src/github.com/klauspost/compress/flate/crc32_amd64.go
… vendor/src/github.com/klauspost/compress/flate/crc32_amd64.s
… vendor/src/github.com/klauspost/compress/flate/crc32_noasm.go
… vendor/src/github.com/klauspost/compress/flate/deflate.go
… vendor/src/github.com/klauspost/compress/flate/fixedhuff.go
… vendor/src/github.com/klauspost/compress/flate/gen.go
… vendor/src/github.com/klauspost/compress/flate/huffman_bit_writer.go
… vendor/src/github.com/klauspost/compress/flate/huffman_code.go
… vendor/src/github.com/klauspost/compress/flate/inflate.go
… vendor/src/github.com/klauspost/compress/flate/reverse_bits.go
… vendor/src/github.com/klauspost/compress/flate/snappy.go
… vendor/src/github.com/klauspost/compress/flate/token.go
… vendor/src/github.com/klauspost/cpuid/.gitignore
… vendor/src/github.com/klauspost/cpuid/.travis.yml
… vendor/src/github.com/klauspost/cpuid/LICENSE
… vendor/src/github.com/klauspost/cpuid/README.md
… vendor/src/github.com/klauspost/cpuid/cpuid.go
… vendor/src/github.com/klauspost/cpuid/cpuid_386.s
… vendor/src/github.com/klauspost/cpuid/cpuid_amd64.s
… vendor/src/github.com/klauspost/cpuid/detect_intel.go
… vendor/src/github.com/klauspost/cpuid/detect_ref.go
… vendor/src/github.com/klauspost/cpuid/generate.go
… vendor/src/github.com/klauspost/cpuid/private-gen.go
… vendor/src/github.com/klauspost/crc32/.gitignore
… vendor/src/github.com/klauspost/crc32/.travis.yml
… vendor/src/github.com/klauspost/crc32/LICENSE
… vendor/src/github.com/klauspost/crc32/README.md
… vendor/src/github.com/klauspost/crc32/crc32.go