/cc @burke 

This replaces the use of the stdlib gzip library with @klauspost's parallel gzip library. More information can be found in this blog post by the author: https://blog.klauspost.com/optimized-gzipzip-packages-30-50-faster/.

This provides a significant speed-up when generating large-ish (1GB or so) container images for the Shopify Rails application. When setting `GOMAXPROCS` to 8, we were observed a 70% performance improvement with `docker push`.

Before:

```
Nov  4 00:39:21.195:    ++ docker push registry.xxx.xxx/build-intermediate-shopify-ci-branches:49ecb981-2b55e03c
Nov  4 00:39:21.296:    The push refers to a repository [registry.xxx.xxx/build-intermediate-shopify-ci-branches] (len: 1)
Nov  4 00:39:21.314:    634c07f5a0c5: Preparing
Nov  4 00:39:21.414:    634c07f5a0c5: Pushing
Nov  4 00:41:19.541:    634c07f5a0c5: Pushed
Nov  4 00:41:19.923:    49ecb98193c3: Image already exists
Nov  4 00:41:20.313:    7411b4ebe7e1: Image already exists
Nov  4 00:41:20.976:    87aaf62b5b12: Image already exists
Nov  4 00:41:21.242:    35c6fdcbcc02: Image already exists
Nov  4 00:41:21.267:    49ecb981-2b55e03c: digest: sha256:1cdc301af17222b34d76e83d603d14bcc0bb5d52f31db6a905630267f23bd270 size: 8557
```

After:

```
Nov  4 00:51:22.912:    ++ docker push registry.xxx.xxx/build-intermediate-shopify-ci-branches:49ecb981-1fa098cf
Nov  4 00:51:23.006:    The push refers to a repository [registry.xxx.xxx/build-intermediate-shopify-ci-branches] (len: 1)
Nov  4 00:51:23.025:    857f975abc73: Preparing
Nov  4 00:51:23.214:    857f975abc73: Pushing
Nov  4 00:52:31.623:    857f975abc73: Pushed
Nov  4 00:52:31.839:    49ecb98193c3: Image already exists
Nov  4 00:52:32.037:    7411b4ebe7e1: Image already exists
Nov  4 00:52:32.381:    87aaf62b5b12: Image already exists
Nov  4 00:52:32.748:    35c6fdcbcc02: Image already exists
Nov  4 00:52:32.750:    49ecb981-1fa098cf: digest: sha256:0a842eb5815fc1715989b4a3ad57ce926bc0edd129231218ec0ca53bc73b250a size: 8557
```