@stevvooe 2015-11-04
This is some good work. klauspost's work is excellent and it will be great to see it in the Go standard library.

My main objections to merging this are the following:
- What is the cost? How much more memory and goroutines are used?
- How does the use of this library affect hash stability of compressed objects?
- How does this perform on other platforms? Many of these optimizations are targeted at amd64.
- How long will klauspost continue to maintain this library? How do we know it is correct? How can we ensure that this dependency is always the most performant? Without the effort to continue this analysis in the future, I'm afraid we'll get stuck on something that is unmaintained, but "faster" because it was at this time.

Optimizations like this are great, but we give up a lot by moving off the standard library. Once we make that decision, we make it even harder to undo.