@klauspost 2015-11-04
@stevvooe - Thanks for the kind words. My plan is to submit the main (compress package)  improvement for Go 1.7. However, I don't think the pgzip package will or should be part of the standard library. The crc32 improvements will be part of Go 1.6. This accounts for ~10% of the total improvement.

A thing I would like to know is the most common data sizes. I tried to make this table, to get an overview of the common data sizes


| File                      | Common Size  | Use pgzip | Use [compress/gzip](https://github.com/klauspost/compress) | Content           |
|---------------------------|--------------|-----------|-------------------|-------------------|
| graph/graph.go            | Small?       |           | x                 | JSON Metadata  only?  |
| graph/push_v2.go          | Small?       |           | x                 | JSON Metadata only?   |
| pkg/archive/archive.go    | Big          | x         |                   | Docker Image Data |
| pkg/tarsum/tarsum.go      | Big variance |           | x                 | Image layer       |
| pkg/tarsum/tarsum_test.go | -            |           | x                 | Must match above  |

I tried to indicate if the best choice would be pgzip or klauspost/compress/gzip. I am a bit unsure about the "tarsum". I assume the payload are rather small, since they are compressed to memory and never a temporary file.

In conclusion, I would recommend using pgzip for the archives and github.com/klauspost/compress/gzip for the "smaller" payloads, unless I have made a mistake in estimating the typical workload sizes.


> What is the cost? How much more memory and goroutines are used?

With default settings 16 blocks of 250000 bytes each can be "in flight". That means a maximum of 8MB buffers, and 16 goroutines (they are reused). The number of in-fligh buffers can be [adjusted](https://godoc.org/github.com/klauspost/pgzip#Writer.SetConcurrency), but there shouldn't be any reason to do so.

[klauspost/compress/gzip](https://github.com/klauspost/compress) uses less memory than the standard library, and has way less allocations.

> How does the use of this library affect hash stability of compressed objects?

Output should remain deterministic for the same version of the library and since you are vendoring, that will not be changing unless you update.

> How does this perform on other platforms? Many of these optimizations are targeted at amd64.
The parallel aspect (pgzip) is platform independent. 

Most of the "compress" optimizations are also platform independent- amd64 is only about 5% faster. You can see [this table](https://github.com/klauspost/compress#deflate-optimizations), where you can see the improvement without assembly below the amd64 improvements.

Since you are compressing disk images, you will gain highly from the increased speed of uncompressible data. Compressing data that is already compressed gives a 2.5x speedup per core on default level compared to the standard library.

>How long will klauspost continue to maintain this library? How do we know it is correct? 
As with most OS packages I cannot make total guarantees neither on support nor correctness. 

I have of course added extensive tests outside the ones in the standard library, but a guarantee for correctness seems a bit much as a requirement for a package. I have kept, and intend to keep updating it with the changes made to the standard library.

>How can we ensure that this dependency is always the most performant? 
I guess you can add benchmarks.

>Without the effort to continue this analysis in the future, I'm afraid we'll get stuck on something that is unmaintained, but "faster" because it was at this time.

That is why the package is compatible with the standard library. If you find yourself "stuck", the only thing you have to do is change the import back to the standard library.