@thaJeztah 2015-11-04
Linking this to https://github.com/docker/docker/issues/10181, which is related

also /cc @unclejack @bobrik @icecrime @aaronlehmann