@unclejack 2015-11-04
`compress/gzip` (the stdlib version) was generating some gzip archives which were making gzip throw errors when reaching the end of the archive. Has this been fixed in the meantime or is it still a problem, both for `compress/gzip` from the stdlib and for the optimized one?

The switch to pgzip will make compression faster, but it will also mean that CPU usage will scale to use up all (if not only most of the CPU cores). This means controls are necessary to avoid having CPU usage go to 100% on all cores.

This change seems premature as `compress/gzip` can still be optimized further. I'd make `compress/gzip` as fast as possible and only look at other options if that's not fast enough.