@csfrancis 2015-11-04
Thank you @klauspost for your comments! I'll try to collect some information from our usage to help better fill out the table that you've created above.

> The switch to pgzip will make compression faster, but it will also mean that CPU usage will scale to use up all (if not only most of the CPU cores). This means controls are necessary to avoid having CPU usage go to 100% on all cores.

Not trying to be a jerk, but isn't that what `GOMAXPROCS` is for?

> This change seems premature as compress/gzip can still be optimized further. I'd make compress/gzip as fast as possible and only look at other options if that's not fast enough.

Of course `compress/gzip` can be further optimized, but that to me seems like a very slow path to getting real results. It will take months/years for the optimizations that klauspost has made to make it into the stdlib, if they even make it at all. The beauty of his library is that it's so simple to drop-in - it only requires modifying the import path. If there are problems, or if performance starts to lag behind the stdlib, it's very simple to remove.